pyproject.toml
[tool.poetry]
name = "projektarbeit_ann"
version = "0.1.0"
description = "Steganalyse mit Deep Learning auf dem ALASKA2 Datensatz"
authors = ["Rino Albertin <rino.albertin@ost.ch>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.1.1"
notebook = "^7.4.1"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
matplotlib = "^3.10.1"
numpy = "^2.2.5"
datasets = "^3.5.1"
torch = { version = "^2.5.1+cu121", source = "pytorch" }
clip-anytorch = "^2.6.0"
faiss-cpu = "^1.11.0"
requests = "^2.32.3"
pillow = "^11.2.1"
seaborn = "^0.13.2"
plotly = "^6.0.1"
scipy = "^1.15.2"
jpegio = "^0.2.8"
opencv-python-headless = "^4.11.0.86"
scikit-image = "^0.25.2"
torchinfo = "^1.8.0"
timm = "^1.0.15"
torchmetrics = "^1.7.3"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
flake8 = "^7.2.0"
basedpyright = "^1.28.4"
complexipy = "^1.2.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
linting = "src.util.poetry.poetry_lint:main"
